fix: move chain to functional flow functions
