+++ src/functional/flow/chain.spec.ts
