+++ src/functional/flow/index.ts
@@ -1,1 +1,2 @@
+export * from './chain';
 export * from './expand';
