+++ src/numbers/quaternion/interpolation.ts
@@ -1,5 +1,6 @@
 import { clamp } from '../../base';
-import { chain, clamp as clampFn } from '../../functional/base';
+import { chain } from '../../functional';
+import { clamp as clampFn } from '../../functional/base';
 
 import { Quaternion } from './quaternion';
 import { add, scale, normalize, dot, divideBy, invertSign } from './operators';
