+++ src/numbers/quaternion/rotate-vec3.ts
@@ -1,5 +1,5 @@
 import { vec3 } from '../../algebra/vector3';
-import { chain } from '../../functional/base';
+import { chain } from '../../functional';
 import { multiplyWith } from '../../functional/numbers/quaternions/multiply-with';
 
 import { conjugate, fromOmegaVec3, Quaternion, xyz } from '../quaternion';
